Implement the this_task() function to get the currently running task.

+++ src/kernel/arch/x86_64/const.h
@@ -51,11 +51,11 @@
 #define CPU_DATA_SIZE_SHIFT     7
 #define CPU_DATA_SIZE           (1 << CPU_DATA_SIZE_SHIFT)
 #define CPU_DATA_BASE           0xc0060000
-#define CPU_DATA(i)             (CPU_DATA_BASE + ((i) << 7))
+#define CPU_DATA(i)             (CPU_DATA_BASE + ((uint64_t)(i) << 7))
 #define CPU_TSS_BASE            (CPU_DATA_BASE + 0)
-#define CPU_TSS(i)              (CPU_TSS_BASE + ((i) << 7))
+#define CPU_TSS(i)              (CPU_TSS_BASE + ((uint64_t)(i) << 7))
 #define CPU_TASK_BASE           (CPU_DATA_BASE + 104)
-#define CPU_TASK(i)             (CPU_TASK_BASE + ((i) << 7))
+#define CPU_TASK(i)             (CPU_TASK_BASE + ((uint64_t)(i) << 7))
 
 /* struct arch_task */
 #define TASK_RP                 0
